Make: use CURDIR in pattern rules to match generated *.d absolute paths
This is benign for normal files because stat doesn't care if the path
is relative or absolute, but if the *.c or *.cu source file is
generated, then it matters whether they match or not.  Specifically,
it can cause a pattern rule to either not match or to match
twice (once as a relative path via explicit prerequisite in the
makefile and again via the *.d depending on abspath).