NS: clean geometric factors (#323)
Store the geometric transform dX_i/dx_j separate from weighting
in qdata, and do not store the symmetric product (formerly called
wBBJ).  We need unweighted transforms for stabilization and
anisotropic diffusion. Condensing into the symmetrized form and
combining with weighting is a natural optimization for Poisson
benchmarks such as the BPs, but is premature (and a poor choice)
for Navier-Stokes.

Developed-by: Valeria Barra <valeria.barra@colorado.edu>
Co-developed-by: Jeremy Thompson <jeremy.thompson@colorado.edu>
Co-developed-by: Jed Brown <jed@jedbrown.org>

* NS: clean geometric factors

* Change qdata[10] and form (w*X_i,j*X_j,i/detJ) in QFunctions

* NS advection: consolidate index notation with loops

* NS advection: consolidate index notation with loops

Thanks-to: Ken Jansen <kenneth.jansen@colorado.edu>

* Unify loop notation and bug fixes

* NS - minor style consistency and doc updates

* NS - Realign inputs

* NS - minor style consistency and doc updates