NS: add support for free slip wall boundary conditions
* -naturalz is removed since it's too specific to cartesian meshes.

* The new interface is -bc_slip_x, -bc_slip_y, and -bc_slip_z.  These
  take a list of face sets.  For the standard box mesh, these are:
  cf. DMPlexCreateCubeMesh_Internal:

  switch (dim) {
  case 2:
    faceMarkerTop    = 3;
    faceMarkerBottom = 1;
    faceMarkerRight  = 2;
    faceMarkerLeft   = 4;
    break;
  case 3:
    faceMarkerBottom = 1;
    faceMarkerTop    = 2;
    faceMarkerFront  = 3;
    faceMarkerBack   = 4;
    faceMarkerRight  = 5;
    faceMarkerLeft   = 6;

One can run a quasi-2D density current model by setting slip
conditions on the ±y faces and making the domain only one element
thick in that dimension.

  ./navierstokes -petscspace_degree 3 -problem density_current -dm_plex_box_faces 32,1,16 -ly 250 -bc_wall 1,2,5,6 -bc_slip_y 3,4 -ts_adapt_monitor -ts_adapt_safety 0.7

Physically, this is like running the experiment sandwiched in a tank
with free-slip walls in the ±y dimensions.

Note: this feature currently depends on a PETSc bug-fix branch that
has not yet been merged to master.