doc: process restrict qualifier as predefined
Breathe and/or Sphinx lacks support for restrict qualifiers, so use
the workaround suggested here to hide the restrict qualifier from
Breathe/Sphinx.

https://github.com/michaeljones/breathe/issues/477#issuecomment-593145041

An alternative is to use the C domain for *.c and *.h files.  This
handles restrict, but reduces enum names to their numeric value.  Once
that Breathe/Sphinx bug is fixed, we could skip the restrict elision
in favor of using the C domain.