pragma vector for Intel; pragma simd is deprecated
Pragma simd is documented without deprecation notice, but the
icc-19.05.281 warns of its deprecation.
https://software.intel.com/en-us/cpp-compiler-developer-guide-and-reference-simdvector

One would think Intel's pragma ivdep would be similar to GCC ivdep,
but icc evidently has different semantics/miscompiles, as demonstrated
in the example, which is left in the source to protect an enterprising
future committer from adding ivdep.