sphinx: prepare for sphinxcontrib-bibtex 2.0, but pin to <2.0
The new sphinxcontrib-bibtex supports parallel building and requires the
listing of all *.bib sources in bibtex_bibfiles, but it also requires
that sphinx be run twice, first to generate a bibtex.json file and then
again to insert the references. I don't know a clean way to do this on
ReadTheDocs so let's stick with 1.x until this issue is resolved or we
need new features.

https://github.com/mcmtroffaes/sphinxcontrib-bibtex/issues/215