examples/fluids: add IJacobian for SUPG
I think this is exact for Galerkin on channel.yaml and blasius.yaml.
Specifically, SNES converges almost perfectly in this test than when
-snes_mf_operator is added.

  -options_file examples/fluids/channel.yaml -snes_monitor -ksp_converged_reason -stab none -ts_max_steps 3 -newtonian_unit_tests -ksp_rtol 1e-10 -order 2 -ts_dt 3e-6 -pc_type lu

I say "think" because something is fishy with the viscous energy term
velocity*stress. If I disable that term, then convergence is clearly
better with the analytic Jacobian (makes sense because it's accuracy is
machine epsilon) than with -snes_mf_operator. This is unexpected, but
the difference is near the limit of numerical stability so I'm not
confident it's wrong. Also, the term itself is so simple.

The Jacobian is inexact for SUPG in that Tau_d is frozen.

WIP: examples/fluids: support MatShell and Blasius_Inflow_Jacobian