t568-operator: fix input order and make test handle nan; skip on broken backends
if (nan < constant) is always false. This test had a memory error
causing it to always compute nan in both the libceed and reference
values. This conditional ensures that output will print in such cases.

This test is BROKEN on /gpu/cuda/gen and /gpu/hip/*. This will need to
be addressed in a future issue.

Reported-by: Natalie Beams <nbeams@icl.utk.edu>