fluids: Fix mass matrix for explicit time stepping
The local vector should be initialized to zero so that all the essential
nodes (ie. strong boundary condition nodes) are not factored in for the
matrix multiply. The PR associated with this commit should have a
document attached detailing why this is the case.

Previous tests used this incorrect mass matrix and so now fail. The
results for these tests are corrected as well.