fluids: Add log_events.c/h, PetscLogEvent for torch
To have the log_events accessbile to torch (in C++), I needed to
separate out the header file containing the extern PetscLogEvent
declarations. While I was at it, I figured it'd be more clear to have a
separate log_events.c file as well to have the actual "storage" of the
PetscLogEvents and the RegisterLogEvents function itself.